most changes complete
